ecourseapisv2/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cloudinary==1.44.1
Django==4.2.28
django-ckeditor==6.7.3
django-js-asset==2.2.0
djangorestframework==3.16.1
PyMySQL==1.1.2
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
